refactor: rename utils folder to types
